requirements.txt
decorator==4.1.2
Django==2.2.5
django-extensions==1.9.6
djangorestframework==3.10.3
Faker==0.8.7
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.6.1
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2019.2
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.3.0
text-unidecode==1.1
traitlets==4.3.2
wcwidth==0.1.7
